Fix potential infinite loop
In finish_compressed_chunk, if the ZlibStream buffer is full but the
input stream is not yet completely consumed then the status will be
HasMoreOutput. This will simply repeat the loop and yet again not
consume more data. We avoid this by transferring some finished data
which leaves at least a full chunk of free space for new data.